+++ test-templates/src/fields.rs
@@ -1,7 +1,7 @@
 #![allow(unused)]
 #![allow(clippy::eq_op)]
 use ark_ff::{
-    fields::{FftField, Field, LegendreSymbol, PrimeField, SquareRootField},
+    fields::{FftField, Field, LegendreSymbol, PrimeField},
     Fp, MontBackend, MontConfig,
 };
 use ark_serialize::{buffer_bit_byte_size, Flags, SWFlags};
@@ -190,7 +190,7 @@ fn random_field_tests<F: Field>() {
     }
 }
 
-fn random_sqrt_tests<F: SquareRootField>() {
+fn random_sqrt_tests<F: Field>() {
     let mut rng = ark_std::test_rng();
 
     for _ in 0..ITERATIONS {
@@ -402,7 +402,7 @@ pub fn montgomery_primefield_test<T: MontConfig<N>, const N: usize>() {
     }
 }
 
-pub fn sqrt_field_test<F: SquareRootField>(elem: F) {
+pub fn sqrt_field_test<F: Field>(elem: F) {
     let square = elem.square();
     let sqrt = square.sqrt().unwrap();
     assert!(sqrt == elem || sqrt == -elem);